moving all internal app code , only kept cli level logic to top directory
