+++ gpt_automation/impl/.gptignore
