+++ gpt_automation/impl/.gptincludeonly
