+++ gpt_automation/impl/base_plugin.py
