+++ gpt_automation/impl/config/.gptincludeonly
