+++ gpt_automation/impl/config/paths.py
