+++ gpt_automation/impl/filter_manager.py
