+++ gpt_automation/impl/ignore_file_parser.py
