+++ gpt_automation/impl/plugin_impl/.gptincludeonly
