+++ gpt_automation/impl/plugins/.gptincludeonly
