+++ gpt_automation/impl/plugins/filter_plugin/manifest.json
