+++ gpt_automation/impl/plugins/filter_plugin/sample_config/black_list.txt
