+++ gpt_automation/impl/plugins/filter_plugin/sample_config/white_list.txt
