+++ gpt_automation/impl/plugins/ignore_plugin/manifest.json
