+++ gpt_automation/impl/plugins/include_only_plugin/manifest.json
