+++ gpt_automation/impl/project_info.py
@@ -1,8 +1,8 @@
 import os
 
-from gpt_automation.config.config_manager import ConfigManager
-from gpt_automation.directory_walker import DirectoryWalker
-from gpt_automation.plugin_impl.manager.plugin_runtime_manager import PluginRuntimeManager
+from gpt_automation.impl.config.config_manager import ConfigManager
+from gpt_automation.impl.directory_walker import DirectoryWalker
+from gpt_automation.impl.plugin_impl.manager.plugin_runtime_manager import PluginRuntimeManager
 
 
 class ProjectInfo:
