+++ gpt_automation/impl/utils/pattern_utils.py
